[Dynamo] Code refactor: move context managers out of misc.py
