+++ torch/_dynamo/variables/__init__.py
@@ -1,6 +1,13 @@
 from .base import VariableTracker
 from .builtin import BuiltinVariable
 from .constant import ConstantVariable, EnumVariable
+from .ctx_manager import (
+    ContextWrappingVariable,
+    CUDAStreamContextVariable,
+    CUDAStreamVariable,
+    GradModeVariable,
+    WithExitFunctionVariable,
+)
 from .dicts import ConstDictVariable, DataClassVariable, DefaultDictVariable
 from .functions import (
     NestedUserFunctionVariable,
@@ -20,12 +27,8 @@ from .misc import (
     AutogradFunctionVariable,
     BlackHoleVariable,
     ClosureVariable,
-    ContextWrappingVariable,
-    CUDAStreamContextVariable,
-    CUDAStreamVariable,
     DeterministicAlgorithmsVariable,
     GetAttrVariable,
-    GradModeVariable,
     InspectSignatureVariable,
     LambdaVariable,
     NewCellVariable,
@@ -34,7 +37,6 @@ from .misc import (
     PythonModuleVariable,
     SuperVariable,
     UnknownVariable,
-    WithExitFunctionVariable,
 )
 from .nn_module import NNModuleVariable, UnspecializedNNModuleVariable
 from .tensor import (
